move the success router and associated files to own dir
