+++ pages/common/routers/index.js
@@ -1,6 +1,5 @@
 module.exports = {
   datatable: require('./datatable'),
   relatedTasks: require('./related-tasks'),
-  form: require('./form'),
-  success: require('./success')
+  form: require('./form')
 };
