+++ pages/pil/dashboard/index.js
@@ -2,7 +2,7 @@ const { page } = require('@asl/service/ui');
 const UnauthorisedError = require('@asl/service/errors/unauthorised');
 const { get, pick, merge, every } = require('lodash');
 const form = require('../../common/routers/form');
-const { success } = require('../../common/routers');
+const success = require('../../success');
 const { hydrate, updateDataFromTask, redirectToTaskIfOpen } = require('../../common/middleware');
 const { saveTaskIdToSession } = require('../../common/helpers');
 const { canUpdateModel, canTransferPil } = require('../../../lib/utils');
