+++ pages/task/read/routes.js
@@ -1,5 +1,5 @@
 const { extend, confirm, read, discard } = require('./routers');
-const success = require('../../common/routers/success');
+const success = require('../../success');
 
 module.exports = {
   root: {
